ml/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
scikit-learn>=1.0.2,<1.4.0
pandas>=1.3.0
numpy>=1.20.0
transformers>=4.18.0
torch>=1.10.0
tensorflow>=2.8.0
nltk==3.8.1
spacy==3.7.2
gensim==4.3.2
python-multipart==0.0.7
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.5